refactor 1/n: carve playback.ts into per-domain modules
src/app/playback.ts (1,706 lines) becomes src/app/playback/ with an
index.ts barrel preserving the exact public surface — no caller or test
changes. Domain modules, in dependency order: status-ui (store-render
sync + modal open/close), wake-lock, transport (timers, stop/pause/
start/toggle), swing (render-token lifecycle), tuning-forms (tuning/
extras forms + deleted-edge/anchor URL state), sleep-timer, track-load
(reset, audio, analysis apply, poll loop, load orchestration, playlist
reconciliation). All functions keep the context: AppContext parameter —
this checkpoint is file moves plus import wiring only.

Judgment call: isPlaybackBlockedForSwing lives in transport (its only
callers) so swing → transport stays one-directional; the import graph
is acyclic: status-ui/wake-lock → transport → swing → tuning-forms →
sleep-timer/track-load.

Gates: typecheck, lint, 498 unit tests, build, e2e 87 passed/1 skipped.
